+++ pyilluminate/illuminate.py
@@ -2,9 +2,11 @@ import json
 import re
 from time import sleep
 from warnings import warn
-from typing import List, Union, Optional, Iterable
+from typing import List, Union, Optional, Iterable, Tuple
 import collections
 from distutils.version import LooseVersion
+from dataclasses import dataclass
+import xarray as xr
 
 import numpy as np
 from serial import Serial, SerialException
@@ -22,55 +24,28 @@ The modes can be:
 """
 
 
-# @dataclass  # python 3.7 only
+@dataclass
 class LEDColor:
     """Generic class for representing colors."""
+    red: int = 0
+    green: int = 0
+    blue: int = 0
+    brightness: Optional[int] = None
 
-    def __init__(self, *, red: int=0, green: int=0, blue: int=0,
-                 brightness: int=None) -> None:
-        """Specify the RGB color of the LEDColor."""
-        if brightness is None:
-            self.red = red
-            self.green = green
-            self.blue = blue
-        else:
-            self.red = brightness
-            self.green = brightness
-            self.blue = brightness
+    def __init__(self, *, brightness: int=None,
+                 red: int=0, green: int=0, blue: int=0):
+        if brightness is not None:
+            red = brightness
+            green = brightness
+            blue = brightness
+        self.brightness = brightness
 
-    def __str__(self):
-        """Print as 'red.green.blue'."""
-        cmd = [str(part) for part in self]
-        return '.'.join(cmd)
-
-    def __repr__(self):
-        return f"LEDColor(red={self.red}, green={self.gree}, blue={self.blue}"
+        self.red = int(red)
+        self.green = int(green)
+        self.blue = int(blue)
 
     def __iter__(self):
-        for i in self.red, self.green, self.blue:
-            yield i
-
-
-class LED:
-    """Generic LED class for setting patterns."""
-
-    def __init__(self, *, led: int=None, red: int=0,
-                 green: int=0, blue: int=0) -> None:
-        """Specify the LED number and the RGB color."""
-        self.led = led
-        self.red = red
-        self.green = green
-        self.blue = blue
-
-    def __str__(self):
-        """Print as "led.red.green.blue"."""
-        if self.led is None:
-            cmd = [self.red, self.green, self.blue]
-        else:
-            cmd = [self.led, self.red, self.green, self.blue]
-
-        cmd = [str(part) for part in cmd]
-        return '.'.join(cmd)
+        return (i for i in (self.red, self.green, self.blue))
 
 
 class Illuminate:
@@ -178,16 +153,21 @@ class Illuminate:
         p = json.loads(self._ask_string('pledpos'))[
             'led_position_list_cartesian']
         self.N_leds = len(p)
-        self._led_positions = np.empty(self.N_leds, dtype=[('x', float),
-                                                           ('y', float),
-                                                           ('z', float)])
 
+        led_positions = xr.DataArray(
+            np.empty((self.N_leds, 3)),
+            dims=['led_number', 'zyx'],
+            coords={'led_number': np.arange(self.N_leds),
+                    'zyx': ['z', 'y', 'x']})
         for key, item in p.items():
-            # x, y units in mm
-            # z unitz in cm
-            self._led_positions[int(key)]['x'] = item[0] * 0.001
-            self._led_positions[int(key)]['y'] = item[1] * 0.001
-            self._led_positions[int(key)]['z'] = item[2] * 0.01
+            # x, y are provided in units of mm
+            # z is provided in units of cm
+            key = int(key)
+            led_positions[key, 2] = item[0] * 0.001
+            led_positions[key, 1] = item[1] * 0.001
+            led_positions[key, 0] = item[2] * 0.01
+
+        self._led_positions = led_positions
 
     @property
     def led_count(self) -> int:
@@ -457,9 +437,14 @@ class Illuminate:
         return self._color
 
     @color.setter
-    def color(self, c: LEDColor):
+    def color(self, c: Union[LEDColor, int, Tuple[int, int, int], List[int]]):
         # sc, [rgbVal] - -or-- sc.[rVal].[gVal].[bVal]
-        self.ask('sc.' + str(c))
+        if isinstance(c, (list, tuple)):
+            c = LEDColor(red=c[0], green=c[1], blue=c[2])
+        elif not isinstance(c, LEDColor):
+            c = LEDColor(brightness=c)
+
+        self.ask(f'sc.{c.red}.{c.green}.{c.blue}')
         self._color = c
 
     @property
@@ -822,7 +807,7 @@ class Illuminate:
         return self._ask_string('pp')
 
     @property
-    def led_positions(self):
+    def led_positions(self) -> xr.DataArray:
         """Position of each LED in cartesian coordinates[mm]."""
         return self._led_positions
 
@@ -835,28 +820,11 @@ class Illuminate:
             This dataarray contains a Nx3 matrix that has rows with the
             ``z, y, x`` coordinates of the leds.
         """
-        import xarray as xr
-        positions = np.empty((len(self.led_positions), 3))
-        positions[:, 2] = self.led_positions['x']
-        positions[:, 1] = self.led_positions['y']
-        positions[:, 0] = self.led_positions['z']
-        # The boards have a 1 mm offset in each dimension I think
-        positions[:, 1:3] += 0.001
-
-        positions = xr.DataArray(
-            positions, dims=['led_number', 'zyx'],
-            coords={'led_number': np.arange(len(positions)),
-                    'zyx': ['z', 'y', 'x']})
-
-        uv_leds = self.uv_leds
-
-        rgb_or_uv = xr.DataArray(
-            np.full(len(positions), fill_value='rgb', dtype='<U3'),
-            dims=['led_number'],
-            coords={'led_number': np.arange(len(positions))})
-        rgb_or_uv[uv_leds] = 'uv'
-        positions = positions.assign_coords(rgb_or_uv=rgb_or_uv)
-        return positions
+        from warnigns import warn
+        warn("The positons_as_xarray function has been Deprecated and will be "
+             "removed in a future version. Use the led_positions attribute "
+             "directly.")
+        return self.led_positions
 
     @property
     def led_positions_NA(self):